requirements.txt
fastapi>=0.68.0
uvicorn>=0.27.0
python-dotenv>=0.19.0
httpx>=0.21.0
pydantic>=1.8.2
supabase>=0.1.10
pytest>=6.2.5
pytest-asyncio>=0.16.0
openai~=1.55.1
cosmpy>=0.9.2  # Add this for XION interactions
